Make program a Flask application
The program (app.py) is now running a Flask app. It responds to
POST-requests on /scanish on the machine it is deployed (ex:
http://localhost:5000/scanish).

It expects url-encoded data where "text" shall be in the the format
specified by the arguments added to the argument parser. Ex: Using curl
--data "text=--scanish vann arbejda" will translate the text to "vatten
arbeta" which is returned in a json element with the same name; "text".

In case of failure, an "error" response; Http response with code 200
with text set to the usage of the app and response_type set to
"ephemeral", meaning that only the user who did the invocation will see
the response.