Integrate with slack message events
Slack events are now picked up by the handleEvents()-method, expecting
HTTP post requests with JSON payload.

200 (OK) response shall always be returned, just to let Slack know the
server recieved the request. The only exception to this is when
validation the URL to the server through Slack, which sends a
"challenge"-value, expecting it in the response.

All user's texts (in the scope; channels as configured in the "My apps"
page at Slack) are now translated to Swedish by default by attaching it
to the original message. This is not done for the invocation of commands
(where the test is prefixed with '/'). Messages with a subtype (as
message_changed, for instance) is not modified as message events are
triggered when the server posts a change.