Add verification of url-encoded requests
Verification failed when Slack sent url-encoded requests (triggered by
slash commands). This was because Flask deserialized the raw body when
request.data was called to log the request. The get_data() method is now
used when logging, and when verifying request, to make sure that the raw
body of the request exists and is used.

Also removed unnecessary import "json".