Parallelize update (translation) of text
Slack expects a response within 3 seconds (as of this writing).
Translating and sending a request to update the text written by a
user may take longer than this.

A new thread is spawned if a user's text shall be translated, which is
responsible for translating and updating the user's text.