Make translation configurable on the user-level
Users can now decide if what they write shall be translated to Swedish
or Scanish by using the new "register"-command. Argument must be
"swedish" or "scanish" or else no translation shall be done. Each time a
user registers themselves, the whole list of users are re-persisted to
disk. This should be fine as users will be few and done seldomly.

Text can no longer be translated through the /scanish context/command;
arguments for this will be split in a future commit so that returned
usage-string will not be confusing.