Reset mocked singletons/static methods after tests for real
Using @classmethod-annotated tearDown()-method and calling reset_mock()
on mock objects only resets the mocks themselves, not the references
used to overwrite them with.

Annotating test-methods with @patch("Class.method") (or as the
documentation says: @patch('module.className')) provides a parameter
(starting from the right...). It has the benefit of restoring the mock:
"The object you specify will be replaced with a mock (or other object)
during the test and restored when the test ends".

For this to work it >seems< the test classes need unittest.TestCase.