Fix command parsing
For some reason, the "unregister"-command is filled with whatever
non-positional argument the user inputs, ex: /scanish anyCommand. This
sets the value of "unregister" to "anyCommand". Added check to verify
its actually set to the value of itself. Seems like a hack, but can't be
arsed to read argument parser documentation.