Change behavior if command-translation
Currently when translating using the command-operation (/Scanish text),
the user's original text is posted to the chat, as well as the
translation by the Scanish app.

It is changed so that only the translation by the Scanish app is posted
to the chat by first returning an empty body and then making a callback
to the response_url received by Slack.

PS: Not even response_type is used in the immediate response since this
will leave the user's input in the input field.

Renamed EventTranslator (to CallbackHandler) since it is now used by the command operation
and not only the event operation. It and the ChatUpdater classes suffer
a bit from copy-paste and can/should be refactored for code re-use.